Welcome Page, Image loads (but Unable to navigate to other pages)
I made a welcome image on Sketch and embedded it on the welcome page. But the Let's get started button is not navigating anywhere, and I am unable to understand the error. I tried to google it, but no luck.

If you guys want to see what the page looks like, the .py file and the image have to be in the same folder.